Resume Review/Project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grog>=0.93",
    "groq>=1.1.2",
    "openai>=2.30.0",
    "pypdf>=6.9.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
